+++ pacman/src/main/java/de/amr/games/pacman/heaven/God.java
@@ -13,7 +13,7 @@ public class God {
 
 	public static final Clock clock = new Clock();
 
-	public static boolean differsAtMost(float value, float target, float tolerance) {
+	public static boolean differsAtMost(double value, double target, double tolerance) {
 		return Math.abs(value - target) <= tolerance;
 	}
 }
